Add InternalAction type for non-public actions.
InternalAction allows Portal to dispatch "special" actions
that cannot be dispatched by Portal's users.

A 'navigateToPreviousRouteAfterBack' internal action is being
added to handle the case of the back transition triggered
by a back button in the navigation bar. Previously a flag
in the 'navigateToPreviousRoute' was needed in order to
differenciate the case where the back transition was
triggered by UIKit or when it was triggered programatically.

This is no longer need thanks to the internal action.

Portal has been refactored in order for other controllers
or components to be able to dispatch internal actions.