Fix issue when rendering change set in controller
New component was not being retained by the controller so future diff
where done against the first component associated with the controller.

To avoid inconsistency issue that may occur if we let users of the
controller API to provide any pair of change set and component they
want, `ComponentPatch` type was added. Only the `PortalViewController`
instances are able to create instances of `ComponentPatch` through
`calculatePatch(for:)` method and in order to update the controller’s
view uses would have to call `render(patch:)`.