Update SwiftLint to version 0.25.0
Fix most new warnings with the exception of colon rule warning
for dictionary type declaration, which I don't agreed with.

Aparently there is a bug in SwiftLint with dictionary that have
Any as the value type.

  https://github.com/realm/SwiftLint/issues/1901

I'll try to deep dive into it and try to submit a patch.
Otherwise I'll ignore this rule temporary in a future commit.