Fix issue with tables not maintaining scroll position.
Table were not maintaining scroll position after re-rendering.
The issue was caused by miss configuration of table view object
to support dynamic size cells.

Table view's estimatedHeight and rowHeight needed to be set.
rowHeight needs to be UITableViewAutomaticDimension and
estimatedHeight, as its name suggest, needs to be the estimated
height of the cells.

In order to calculate the estimated height and because
cell items could have different heights, the most probable
cell heights is computed based on the occurencies of the heights
in the current items set. The height with the most occurrencies
is the estimated height because it has higher change of being
the actual height.